changed acceptable content length type from int to long. Added debug for exceptions

+++ src/main/java/anhttpserver/SimpleHttpHandler.java
@@ -109,7 +109,7 @@ public interface SimpleHttpHandler {
      *  facade for {@link com.sun.net.httpserver.HttpExchange}
      * @return response size
      */
-    public int getResponseSize(HttpRequestContext httpRequestContext);
+    public long getResponseSize(HttpRequestContext httpRequestContext);
 
     /**
      * Somewhy in {@link com.sun.net.httpserver.HttpExchange} context attibutes
